Added Card component
Data rendering for each monster is now passed to that component.
CardList is responsible for layout of cards and iterating monsters
to pass each one to a Card child component.

Also pay attention to CardList component giving each card a unique key attribute.
This is very important for performance when react needs to determine what changed and
change only the affected components. Component lists like these are the ones responsible for
giving out keys to individual children, not children themselves.