Turning onChange event handler into a class method.
Neat trick - making class method an arrow function binds this in it perma to the
class instance object. Since arrow functions have no "this" of their own, it is taken from
the nearest parent that has it. Unlike regular functions whose this is defined by method of invocation,
arrow function's is defined by place of declaration.

With this we can skip .bind(this) calls. Neat!